# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Online Event Booking API
  description: Tài liệu OpenAPI / Swagger hoàn chỉnh cho toàn bộ hệ thống Online Event Booking, tích hợp xác thực JWT và Trợ lý AI.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local Development Server (với context-path /api)
...
paths:
...
  /auth/token:
    post:
      tags:
        - Authentication
      summary: Đăng nhập lấy JWT Token
      description: Xác thực tài khoản bằng email và mật khẩu để nhận Access Token.
      operationId: authenticate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationRequest'
      responses:
        '200':
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuthenticationResponse'
  /auth/introspect:
    post:
      tags:
        - Authentication
      summary: Kiểm tra tính hợp lệ của Token
      description: Kiểm tra xem JWT Token có hợp lệ hoặc đã bị vô hiệu hóa (logout) hay chưa.
      operationId: introspect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntrospectRequest'
      responses:
        '200':
          description: Kết quả kiểm tra token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseIntrospectResponse'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Đăng xuất tài khoản
      description: Vô hiệu hóa JWT Token hiện tại và lưu vào danh sách Blacklist.
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Đăng xuất thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'

  # ================= ROLES =================
...
  /permissions:
    post:
      tags:
        - Permission Management
      summary: Tạo quyền hạn mới
      description: Tạo mới một quyền hạn trong hệ thống (Yêu cầu ADMIN).
      operationId: createPermission
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionRequest'
      responses:
        '200':
          description: Tạo quyền hạn thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePermissionResponse'
    get:
      tags:
        - Permission Management
      summary: Lấy danh sách tất cả quyền hạn
      description: Lấy danh sách toàn bộ quyền hạn hiện có (Yêu cầu ADMIN).
      operationId: getAllPermissions
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Danh sách quyền hạn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseListPermissionResponse'
  /permissions/{permission}:
    delete:
      tags:
        - Permission Management
      summary: Xóa một quyền hạn
      description: Xóa quyền hạn theo tên mã (Yêu cầu ADMIN).
      operationId: deletePermission
      security:
        - BearerAuth: []
      parameters:
        - name: permission
          in: path
          required: true
          description: Tên quyền hạn cần xóa (ví dụ CREATE_EVENT)
          schema:
            type: string
      responses:
        '200':
          description: Xóa quyền hạn thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'

  # ================= CUSTOMERS =================
  /customers:
    post:
      tags:
        - Customer Management
      summary: Đăng ký tài khoản khách hàng mới
      description: Tạo mới tài khoản khách hàng với thông tin cơ bản.
      operationId: registerCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '200':
          description: Đăng ký thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCustomerResponse'
  /customers/{id}/bookings:
    get:
      tags:
        - Customer Management
      summary: Xem lịch sử đặt vé của khách hàng
      description: Lấy danh sách toàn bộ các lượt đặt chỗ của khách hàng theo ID.
      operationId: getCustomerBookings
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID khách hàng
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Danh sách đặt vé của khách hàng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseListCustomerBookingsResponse'

  # ================= EVENTS =================
  /events:
    post:
      tags:
        - Event Management
      summary: Tạo sự kiện mới và loại vé đi kèm
      description: Tạo sự kiện và các loại vé liên kết trong cùng một giao dịch (Yêu cầu ADMIN hoặc quyền CREATE_EVENT).
      operationId: createEvent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
      responses:
        '200':
          description: Tạo sự kiện thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEventResponse'
    get:
      tags:
        - Event Management
      summary: Tìm kiếm và phân trang sự kiện
      description: Hỗ trợ tìm kiếm theo tiêu đề, thể loại, địa điểm, khoảng thời gian kết hợp phân trang và sắp xếp.
      operationId: getEvents
      parameters:
        - name: title
          in: query
          required: false
          schema:
            type: string
          description: Tiêu đề sự kiện cần tìm kiếm
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: Thể loại sự kiện (MUSIC, TECH, FOOD, ART...)
        - name: venue
          in: query
          required: false
          schema:
            type: string
          description: Địa điểm tổ chức
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Thời gian bắt đầu (từ thời điểm này trở đi)
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Thời gian kết thúc
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Số thứ tự trang (bắt đầu từ 0)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: Số lượng sự kiện trên một trang
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            default: id
          description: Trường sắp xếp (id, startTime, title...)
        - name: sortDir
          in: query
          required: false
          schema:
            type: string
            default: asc
          description: Hướng sắp xếp (asc hoặc desc)
      responses:
        '200':
          description: Lấy danh sách sự kiện thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageEventResponse'
  /events/{id}:
    get:
      tags:
        - Event Management
      summary: Xem chi tiết sự kiện theo ID
      description: Lấy thông tin chi tiết sự kiện cùng danh sách loại vé còn lại.
      operationId: getEventById
      parameters:
        - name: id
          in: path
          required: true
          description: ID của sự kiện
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lấy chi tiết thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEventResponse'
  /events/{id}/ask:
    post:
      tags:
        - AI Assistant
      summary: Hỏi đáp AI về một sự kiện cụ thể
      description: Trợ lý AI giải đáp thắc mắc của người dùng dựa trên thông tin chi tiết của sự kiện.
      operationId: askEventIdWithAi
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID sự kiện
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiRequest'
      responses:
        '200':
          description: AI trả lời thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAiResponse'

  # ================= BOOKINGS =================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Nhập JWT Token vào đây (không cần tiền tố Bearer)"
  schemas:
...
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: integer
          example: 1000
        message:
          type: string
          example: "Thành công"
    ApiResponseAuthenticationResponse:
      type: object
      properties:
        code:
          type: integer
          example: 1000
        message:
          type: string
          example: "Thành công"
        result:
          $ref: '#/components/schemas/AuthenticationResponse'
    ApiResponseIntrospectResponse:
      type: object
      properties:
        code:
          type: integer
          example: 1000
        message:
          type: string
          example: "Thành công"
        result:
          $ref: '#/components/schemas/IntrospectResponse'
...
    ApiResponsePermissionResponse:
      type: object
      properties:
        code:
          type: integer
          example: 1000
        message:
          type: string
          example: "Thành công"
        result:
          $ref: '#/components/schemas/PermissionResponse'
    ApiResponseListPermissionResponse:
      type: object
      properties:
        code:
          type: integer
          example: 1000
        message:
          type: string
          example: "Thành công"
        result:
          type: array
          items:
            $ref: '#/components/schemas/PermissionResponse'
    ApiResponseCustomerResponse:
      type: object
      properties:
        code:
          type: integer
          example: 1000
        message:
          type: string
          example: "Thành công"
        result:
          $ref: '#/components/schemas/CustomerResponse'
    ApiResponseListCustomerBookingsResponse:
      type: object
      properties:
        code:
          type: integer
          example: 1000
        message:
          type: string
          example: "Thành công"
        result:
          type: array
          items:
            $ref: '#/components/schemas/CustomerBookingsResponse'
    ApiResponseEventResponse:
      type: object
      properties:
        code:
          type: integer
          example: 1000
        message:
          type: string
          example: "Thành công"
        result:
          $ref: '#/components/schemas/EventResponse'
    ApiResponsePageEventResponse:
      type: object
      properties:
        code:
          type: integer
          example: 1000
        message:
          type: string
          example: "Thành công"
        result:
          $ref: '#/components/schemas/PageEventResponse'
...
    ApiResponseAiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 1000
        message:
          type: string
          example: "Thành công"
        result:
          type: object
          description: Kết quả phản hồi từ AI (JSON Object hoặc String)

    # Auth DTOs
    AuthenticationRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: admin@bookingevent.com
        password:
          type: string
          format: password
          example: admin123
    AuthenticationResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzUxMiJ9..."
        authenticated:
          type: boolean
          example: true
    IntrospectRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzUxMiJ9..."
    IntrospectResponse:
      type: object
      properties:
        valid:
          type: boolean
          example: true
    LogoutRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzUxMiJ9..."

    # Role & Permission DTOs
...
    PermissionRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "CREATE_EVENT"
        description:
          type: string
          example: "Quyền tạo mới sự kiện và loại vé"
    PermissionResponse:
      type: object
      properties:
        name:
          type: string
          example: "CREATE_EVENT"
        description:
          type: string
          example: "Quyền tạo mới sự kiện và loại vé"

    # Customer DTOs
    CustomerRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          example: "Nguyen Van A"
        email:
          type: string
          format: email
          example: "nguyenvana@gmail.com"
        password:
          type: string
          format: password
          example: "12345678"
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Nguyen Van A"
        email:
          type: string
          example: "nguyenvana@gmail.com"
    CustomerBookingsResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        reference:
          type: string
          example: "BK-20260729-0001"
        status:
          type: string
          example: "CONFIRMED"
        totalAmount:
          type: number
          example: 500000.00
        createdAt:
          type: string
          format: date-time
          example: "2026-08-28T09:30:00Z"
        items:
          type: array
          items:
            $ref: '#/components/schemas/CustomerBookingItemResponse'
    CustomerBookingItemResponse:
      type: object
      properties:
        ticketTypeId:
          type: integer
          format: int64
          example: 2
        ticketTypeName:
          type: string
          example: "VIP Ticket"
        quantity:
          type: integer
          example: 2
        unitPrice:
          type: number
          example: 250000.00

    # Event DTOs
    EventRequest:
      type: object
      required:
        - title
        - category
        - venue
        - startTime
        - organizer
        - ticketTypes
      properties:
        title:
          type: string
          example: "Music Concert 2026"
        description:
          type: string
          example: "Đại nhạc hội Pop mùa hè hoành tráng"
        category:
          type: string
          example: "MUSIC"
        venue:
          type: string
          example: "Sân vận động Mỹ Đình"
        startTime:
          type: string
          format: date-time
          example: "2026-08-15T19:30:00Z"
        organizer:
          type: string
          example: "V-Pop Center"
        ticketTypes:
          type: array
          items:
            $ref: '#/components/schemas/TicketRequest'
    TicketRequest:
      type: object
      required:
        - name
        - price
        - totalQuantity
      properties:
        name:
          type: string
          example: "VIP Ticket"
        price:
          type: number
          format: double
          example: 1500000.00
        totalQuantity:
          type: integer
          example: 100
    EventResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: "Music Concert 2026"
        description:
          type: string
          example: "Đại nhạc hội Pop mùa hè hoành tráng"
        category:
          type: string
          example: "MUSIC"
        venue:
          type: string
          example: "Sân vận động Mỹ Đình"
        startTime:
          type: string
          format: date-time
          example: "2026-08-15T19:30:00Z"
        organizer:
          type: string
          example: "V-Pop Center"
        ticketTypes:
          type: array
          items:
            $ref: '#/components/schemas/TicketResponse'
    TicketResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        eventId:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "VIP Ticket"
        price:
          type: number
          format: double
          example: 1500000.00
        totalQuantity:
          type: integer
          example: 100
        soldQuantity:
          type: integer
          example: 0
    PageEventResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/EventResponse'
        pageable:
          $ref: '#/components/schemas/PageableObject'
        totalPages:
          type: integer
          example: 3
        totalElements:
          type: integer
          example: 11
        last:
          type: boolean
          example: false
        size:
          type: integer
          example: 5
        number:
          type: integer
          example: 0
        sort:
          $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          example: 5
        first:
          type: boolean
          example: true
        empty:
          type: boolean
          example: false
    PageableObject:
      type: object
      properties:
        pageNumber:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 5
        sort:
          $ref: '#/components/schemas/SortObject'
        offset:
          type: integer
          example: 0
        paged:
          type: boolean
          example: true
        unpaged:
          type: boolean
          example: false
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
          example: false
        sorted:
          type: boolean
          example: true
        unsorted:
          type: boolean
          example: false

    # Booking DTOs
...
    AiRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: "Nội dung câu hỏi hoặc yêu cầu gửi AI"
          example: "Tìm giúp tôi sự kiện âm nhạc diễn ra vào tháng 8"
        customerId:
          type: integer
          format: int64
          description: "ID khách hàng (nếu đã đăng nhập)"
          example: 1